convert all datapoint to float to avoid converison slowness and overflow

+++ src/game/action.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone)]
 pub struct Action {
-    pub thrust: u8,
-    pub angle: i16,
+    pub thrust: i32,
+    pub angle: i32,
 }
 
 impl Action {
-    pub fn new(thrust: u8, angle: i16) -> Self {
+    pub fn new(thrust: i32, angle: i32) -> Self {
         Self { thrust, angle }
     }
 }
@@ -19,8 +19,8 @@ impl std::fmt::Display for Action {
 impl From<&str> for Action {
     fn from(s: &str) -> Self {
         let mut parts = s.split(',');
-        let thrust = parts.next().unwrap_or("0").parse::<u8>().unwrap_or(0);
-        let angle = parts.next().unwrap_or("0").parse::<i16>().unwrap_or(0);
+        let thrust = parts.next().unwrap_or("0").parse::<i32>().unwrap_or(0);
+        let angle = parts.next().unwrap_or("0").parse::<i32>().unwrap_or(0);
         Action::new(thrust, angle)
     }
 }
